requirements.txt
autopep8==1.4.3
paho-mqtt==1.4.0
pycodestyle==2.4.0
